refactor: user groups to teams, rename frontend files
