+++ public/app/features/org/partials/team_details.html
