+++ public/sass/base/_icons.scss
@@ -52,7 +52,7 @@
 }
 
 .gicon-team {
-  background-image: url('../img/icons_#{$theme-name}_theme/icon_user_group.svg');
+  background-image: url('../img/icons_#{$theme-name}_theme/icon_team.svg');
 }
 
 .gicon-org {
